Refactor
Moved similar functions to a mixin
