+++ vue/components/organisms/ripe-configurator/ripe-configurator.vue
@@ -59,7 +59,6 @@
 </style>
 
 <script>
-import { Ripe, ripe } from "ripe-sdk";
 import { ripe } from "ripe-sdk";
 import { logicMixin } from "../../../mixins";
 import "ripe-sdk/src/css/ripe.css";
@@ -385,42 +384,6 @@ export const RipeConfigurator = {
         this.resize(this.size);
     },
     methods: {
-        /**
-         * Initializes RIPE instance if it does not exists and
-         * configures it with the given brand, model, version
-         * and parts. If a RIPE instance is provided, it will
-         * be used without further configuration.
-         */
-        async setupRipe() {
-            if (!this.ripeData) {
-                this.ripeData = new Ripe();
-            }
-
-            await this.configRipe();
-
-            // in case the global RIPE instance is not set then
-            // updates it with the current one
-            if (!global.ripe) {
-                global.ripe = this.ripeData;
-            }
-        },
-        /**
-         * Configures the RIPE instance with the given brand,
-         * model, version and parts.
-         */
-        async configRipe() {
-            this.loading = true;
-
-            try {
-                await this.ripeData.config(this.brand, this.model, {
-                    version: this.version,
-                    parts: this.partsData
-                });
-            } catch (error) {
-                this.loading = false;
-                throw error;
-            }
-        },
         /**
          * Re-sizes the configurator according to the current
          * available container size (defined by parent).
