+++ vue/mixins/index.js
@@ -1,1 +1,2 @@
 export * from "./money";
+export * from "./logic";
